requirements.txt
python-dotenv==1.0.1
fastapi==0.115.3
mysql-connector-python==9.1.0
uvicorn==0.32.0
python-multipart==0.0.15
jinja2==3.1.4
pydantic[email]
requests==2.32.3